requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
requests
structlog
python-json-logger
prometheus-client
httpx
